--- a/reach/airflow/dags/policy.py
+++ b/reach/airflow/dags/policy.py
@@ -43,7 +43,7 @@ def verify_s3_prefix():
     assert reach_s3_prefix.startswith('s3://')
     assert not reach_s3_prefix.endswith('/')
 
-
+# Force verification of S3 prefix on import
 verify_s3_prefix()
 
 
@@ -74,13 +74,30 @@ def to_s3_output_dir(dag, *args):
     ) % (dag.dag_id, path, slug)
 
 
-def create_org_pipeline(dag, organisation, item_limits, spider_years):
-    """ Creates all tasks tied to a single organisation::
+EPMC_METADATA_KEY = '/'.join([
+    '{{ conf.get("core", "openresearch_s3_prefix") }}',
+    'output', 'open-research', 'epmc-metadata', 'epmc-metadata.json.gz'
+    ])
 
-        Spider -> ParsePdf
-                    \-> ExtractRefs -> FuzzyMatchRefs
-                    \-> ESIndexFulltextdocs -> FullTextMatchRefs
-    """
+
+#
+# Tasks within the DAG
+#
+
+def es_index_publications(dag, item_limits):
+    """ Returns the task for indexing EPMC publications into ES. """
+    return es_index_epmc_metadata.ESIndexEPMCMetadata(
+            task_id='ESIndexEPMCMetadata',
+            src_s3_key=EPMC_METADATA_KEY,
+            es_hosts=get_es_hosts(),
+            max_epmc_metadata=item_limits.index,
+            es_index='-'.join([dag.dag_id, 'epmc', 'metadata']),
+            dag=dag
+        )
+
+def org_refs(dag, organisation, item_limits, spider_years):
+    """ Returns the portion of an organization's pipeline that produces
+    references. """
     spider = SpiderOperator(
         task_id='Spider.%s' % organisation,
         organisation=organisation,
@@ -98,16 +115,13 @@ def create_org_pipeline(dag, organisation, item_limits, spider_years):
             dag, 'parsed-pdfs', organisation, '.json.gz'),
         dag=dag)
 
-    esIndexFullTexts = es_index_fulltext_docs.ESIndexFulltextDocs(
-        task_id="ESIndexFulltextDocs.%s" % organisation,
-        src_s3_key=parsePdf.dst_s3_key,
-        organisation=organisation,
-        es_hosts=get_es_hosts(),
-        item_limits=item_limits.index,
-        es_index='-'.join([dag.dag_id, 'docs']),
-        dag=dag
-    )
+    spider >> parsePdf
+    return parsePdf
 
+
+def org_fuzzy_match(dag, organisation, item_limits, parsePdf, esIndexPublications):
+    """ Returns the fuzzy match portion of the pipeline for a single
+    organization. """
     extractRefs = ExtractRefsOperator(
         task_id='ExtractRefs.%s' % organisation,
         src_s3_key=parsePdf.dst_s3_key,
@@ -135,34 +149,24 @@ def create_org_pipeline(dag, organisation, item_limits, spider_years):
         dag=dag
     )
 
-    parsePdf >> esIndexFullTexts
-    spider >> parsePdf >> extractRefs >> fuzzyMatchRefs >> esIndexFuzzyMatched
-    return fuzzyMatchRefs
+    esIndexPublications >> fuzzyMatchRefs
+    parsePdf >> extractRefs >> fuzzyMatchRefs >> esIndexFuzzyMatched
+    return esIndexFuzzyMatched
 
-def create_org_pipeline_exact_match(dag, organisation, item_limits, spider_years, epmc_metadata_key=None):
-    """ Creates all tasks tied to a single organisation::
+
+def create_org_pipeline_fuzzy_match(dag, organisation, item_limits, spider_years,
+                        esIndexPublications):
+    """ Creates all tasks needed for producing fuzzy matched citations
+    for a single organisation::
 
         Spider -> ParsePdf
-                    \-> ExtractRefs -> FuzzyMatchRefs
-                    \-> ESIndexFulltextdocs -> FullTextMatchRefs
+                    \-> ExtractRefs -> FuzzyMatchRefs -> EsIndexFullTextDocs
+                                         ^
+                                         |
+                                       EsIndexPublications 
     """
-    spider = SpiderOperator(
-        task_id='Spider.%s' % organisation,
-        organisation=organisation,
-        dst_s3_dir=to_s3_output_dir(
-            dag, 'spider', organisation),
-        item_years=spider_years,
-        item_max=item_limits.spiders,
-        dag=dag)
 
-    parsePdf = ParsePdfOperator(
-        task_id='ParsePdf.%s' % organisation,
-        organisation=organisation,
-        src_s3_dir=spider.dst_s3_dir,
-        dst_s3_key=to_s3_output(
-            dag, 'parsed-pdfs', organisation, '.json.gz'),
-        dag=dag)
-
+    parsePdf = org_refs(dag, organisation, item_limits, spider_years)
     esIndexFullTexts = es_index_fulltext_docs.ESIndexFulltextDocs(
         task_id="ESIndexFulltextDocs.%s" % organisation,
         src_s3_key=parsePdf.dst_s3_key,
@@ -172,115 +176,116 @@ def create_org_pipeline_exact_match(dag, organisation, item_limits, spider_years
         es_index='-'.join([dag.dag_id, 'docs']),
         dag=dag
     )
+    parsePdf >> esIndexFullTexts
+    return parsePdf, esIndexFullTexts, org_fuzzy_match(
+        dag, organisation, item_limits, parsePdf, esIndexPublications)
+
+
+def create_org_pipeline_exact_match(dag, organisation, item_limits,
+                                    spider_years, parsePdf=None,
+                                    esIndexFullTexts=None):
+    """ Creates all tasks needed for producing exact matched citatinos
+    for a single organisation::
+
+        Spider -> ParsePdf
+                    \-> ESIndexFulltextdocs -> FullTextMatchRefs
+    """
+
+    if parsePdf is None:
+        parsePdf = org_refs(dag, organisation, item_limits, spider_years)
+
+    if esIndexFullTexts is None:
+        esIndexFullTexts = es_index_fulltext_docs.ESIndexFulltextDocs(
+            task_id="ESIndexFulltextDocs.%s" % organisation,
+            src_s3_key=parsePdf.dst_s3_key,
+            organisation=organisation,
+            es_hosts=get_es_hosts(),
+            item_limits=item_limits.index,
+            es_index='-'.join([dag.dag_id, 'docs']),
+            dag=dag
+        )
 
     exactMatchRefs = exact_match_refs_operator.ExactMatchRefsOperator(
         task_id='ExactMatchRefs.%s' % organisation,
         es_hosts=get_es_hosts(),
-        epmc_metadata_key=epmc_metadata_key,
+        epmc_metadata_key=EPMC_METADATA_KEY,
         exact_matched_references_path=to_s3_output(
             dag, 'exact-matched-refs', organisation, '.json.gz'),
+        item_limits=item_limits.index,
         es_full_text_index='-'.join([dag.dag_id, 'docs']),
         dag=dag)
-
-
-    spider >> parsePdf >> esIndexFullTexts >> exactMatchRefs
+    parsePdf >> esIndexFullTexts >> exactMatchRefs
     return exactMatchRefs
 
-def create_dag(dag_id, default_args, spider_years,
+
+def create_dag_fuzzy_match(dag_id, default_args, spider_years,
                item_limits):
     """
     Creates a DAG.
 
     Args:
+        dag_id: dag id
         default_args: default args for the DAG
-        spider_op_cls: Spider operator class.
+        spider_years: years to scrape
+        item_limits: ItemLimits instance to limit number of items
     """
     dag = DAG(
         dag_id=dag_id,
         default_args=default_args,
         schedule_interval='0 0 * * 0,3'
     )
-
-    epmc_metadata_key = '/'.join([
-        '{{ conf.get("core", "openresearch_s3_prefix") }}',
-        'output', 'open-research', 'epmc-metadata', 'epmc-metadata.json.gz'
-    ])
-
-    esIndexPublications = es_index_epmc_metadata.ESIndexEPMCMetadata(
-        task_id='ESIndexEPMCMetadata',
-        src_s3_key=epmc_metadata_key,
-        es_hosts=get_es_hosts(),
-        max_epmc_metadata=item_limits.index,
-        es_index='-'.join([dag_id, 'epmc', 'metadata']),
-        dag=dag
-    )
+    esIndexPublications = es_index_publications(dag, item_limits)
     for organisation in ORGANISATIONS:
-        fuzzyMatchRefs = create_org_pipeline(
-            dag,
-            organisation,
-            item_limits,
-            spider_years,
-        )
-        esIndexPublications >> fuzzyMatchRefs
-
+        create_org_pipeline_fuzzy_match(
+            dag, organisation, item_limits, spider_years,
+            esIndexPublications)
     return dag
 
-def create_dag_exact_match(dag_id, default_args, spider_years,
-               item_limits):
+
+def create_dag_all_match(dag_id, default_args, spider_years,
+                         item_limits):
     """
-    Creates a DAG.
+    Creates a DAG that produces citations using both fuzzy and exact
+    matchers.
 
     Args:
+        dag_id: dag id
         default_args: default args for the DAG
-        spider_op_cls: Spider operator class.
+        spider_years: years to scrape
+        item_limits: ItemLimits instance to limit number of items
     """
     dag = DAG(
         dag_id=dag_id,
         default_args=default_args,
         schedule_interval='0 0 * * 0,3'
     )
-
-    epmc_metadata_key = '/'.join([
-        '{{ conf.get("core", "openresearch_s3_prefix") }}',
-        'output', 'open-research', 'epmc-metadata', 'epmc-metadata.json.gz'
-    ])
-
-    esIndexPublications = es_index_epmc_metadata.ESIndexEPMCMetadata(
-        task_id='ESIndexEPMCMetadata',
-        src_s3_key=epmc_metadata_key,
-        es_hosts=get_es_hosts(),
-        max_epmc_metadata=item_limits.index,
-        es_index='-'.join([dag_id, 'epmc', 'metadata']),
-        dag=dag
-    )
+    esIndexPublications = es_index_publications(dag, item_limits)
     for organisation in ORGANISATIONS:
-        exactMatchRefs = create_org_pipeline_exact_match(
+        parsePdf, esIndexFullTexts, _ = create_org_pipeline_fuzzy_match(
             dag,
             organisation,
             item_limits,
             spider_years,
-            epmc_metadata_key,
-        )
-        esIndexPublications >> exactMatchRefs
+            esIndexPublications)
+        create_org_pipeline_exact_match(
+            dag,
+            organisation,
+            item_limits,
+            spider_years,
+            parsePdf=parsePdf,
+            esIndexFullTexts=esIndexFullTexts)
 
     return dag
 
 
-test_dag = create_dag(
+test_dag = create_dag_all_match(
     'policy-test',
     DEFAULT_ARGS,
     [2018],
     ItemLimits(10, 500),
 )
 
-test_dag_exact_match = create_dag_exact_match(
-    'policy-test-exact-match',
-    DEFAULT_ARGS,
-    [2018],
-    ItemLimits(10, 100),
-)
-
-policy_dag = create_dag(
+policy_dag = create_dag_fuzzy_match(
     'policy',
     DEFAULT_ARGS,
     list(range(2012, datetime.datetime.now().year + 1)),