--- a/reach/airflow/tasks/es_index_epmc_metadata.py
+++ b/reach/airflow/tasks/es_index_epmc_metadata.py
@@ -48,7 +48,7 @@ class ESIndexEPMCMetadata(BaseOperator):
 
     def execute(self, context):
         es = reach.elastic.common.connect(
-            self.es_hosts, self.es_port)
+            self.es_hosts)
         s3 = WellcomeS3Hook()
 
         # TODO: implement skipping mechanism