@hblanks 2019-11-19
@nsorros , @lizgzil , @SamDepardieu , @jdu  - OK! I've updated so that:

1. The test DAG has both fuzzy and exact match, with appropriately limited inputs all the way down
1. The main DAG only has fuzzy match

And the DAG generation code is a little bit more modular along the way

This gets us to where at least the test DAG runs in short time, and we can move exact into the main DAG when we speed it up enough.

PTAL when you can, and we'll give it a merge.