@hblanks 2019-11-20
One more update: this does increase the test DAG runtime by ~6 minutes:

![gantt](https://user-images.githubusercontent.com/991383/69248274-e94e0280-0ba3-11ea-9681-03e07dc75e1f.png)

@SamDepardieu , @jdu  - do let me know if you think a 15 minute runtime is acceptable. If not, we still have some options:

1. Reduce number of records matched by exact matcher in test DAG.
1. Increase parallelism (though this may be RAM intensive) by increasing `celery.worker_concurrency` from 2 to 4 in `airflow.cfg`.