@jdu 2019-11-21
15 minutes is fine as i've taken to running a full policy-test and then re-running individual tasks within it to test changes, etc... so it's not all that bad.

Maybe further down the road we can set up a `python -m HTTPServer` based site in the repo or something to have a more controlled/faster test target for the policy-test.